% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2022_12_15-MODULO1-user_id.R
\docType{data}
\name{gtdbK_file}
\alias{gtdbK_file}
\title{Subsets the mapped KO table based on pathways.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 42 rows and 19 columns.
}
\usage{
user_id()
}
\description{
reads the output of the mapping_ko function to filter out KOs
based on pathways.
}
\details{
This function is part of a package used for the analysis of
bins metabolism.
}
\examples{
user_id()
}
\keyword{datasets}
