% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_EvoFiles.R
\name{make_EvoFiles}
\alias{make_EvoFiles}
\title{Make EvoMining files}
\usage{
make_EvoFiles(annotation_dirpath,
                         genome_dirpath,
                             IDs_table)
}
\arguments{
\item{annotation_dirpath}{is the path of annotation output directory created
by [KofamScan](https://github.com/takaram/kofam_scan.git).}

\item{genome_dirpath}{is the path of all protein sequences files of all the
bins}

\item{IDs_table}{is a table of IDs made by make_annotationIDs_bytaxa function}
}
\value{
a list of all sequence IDs in the file
}
\description{
Esta funcion toma El directorio de secuencias de Proteomas
obtenidas de prodigal, el directorio de los outputs de KO y el archivo
resultado de la funcion allrast_names_table y regresa un archivo anotado con
especie y un fasta con los feature ids de rast simulados
}
\details{
This function is part of the MetaEvoMining package
}
\examples{
make_EvoFiles("Datos/02.KO_results_all/",
                        'data/03/03.Proteome_named_scaff/',
                        "Datos/rast_namestable.IDs")
}
