% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_anotationIDs_bytaxa.R
\name{make_rastIDs_by_taxa}
\alias{make_rastIDs_by_taxa}
\title{Make rast IDs file by taxa}
\usage{
make_rastIDs_by_taxa(gtdbK_report,taxa)
}
\arguments{
\item{gtdbK_report}{is a tsv file wich was create by
[gtdbtk](https://github.com/Ecogenomics/GTDBTk.git) program.}

\item{taxa}{is the taxonomic rank}
}
\description{
Esta función toma el archivo gtdb-tk y busca los MAGs que
pertenecen a la misma clase taxonómica y regresa un archivo parecido al ids
de rast las filas son: nombre numérico, identificador único y especie
}
\details{
This function is part of the MODULE2 of MetaEvoMining package
2023_03_15
}
\examples{
make_rastIDs_by_taxa('data/gtdbtk.bac120.summary.tsv',
"f__Saccharospirillaceae")
}
