% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2023_03_15_MODULE2_make_allnames_table.R
\name{make_allnames_table}
\alias{make_allnames_table}
\title{Make created names for evomining}
\usage{
make_allnames_table(gtdbK_report)
}
\arguments{
\item{gtdbK_report}{is a tsv file wich was create by
[gtdbtk](https://github.com/Ecogenomics/GTDBTk.git) program.}
}
\description{
This function reads the gtdb-tk taxonomic mapping summary file and
returns a four-column file containing a 6-digit numeric name, a unique
identifier, the user-assigned genome name, and the taxonomic mapping.
}
\details{
This function is part of the MetaEvoMining package
}
\examples{
make_allnames_table('data/gtdbtk.bac120.summary.tsv')
}
