% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Evo_Files.R
\name{make_EvoFiles}
\alias{make_EvoFiles}
\title{Make EvoMining files}
\usage{
make_EvoFiles(gtdbK_report, genomes_dirpath, annotation_dirpath)
}
\arguments{
\item{gtdbK_report}{gtdbK_report is a tsv file wich was create by
[gtdbtk](https://github.com/Ecogenomics/GTDBTk.git) program.}

\item{genomes_dirpath}{is the path of all protein sequences files of all the
bins}

\item{annotation_dirpath}{is the path of annotation output directory created
by [KofamScan](https://github.com/takaram/kofam_scan.git).}
}
\value{
a list of all sequence IDs in the file
}
\description{
Esta funcion toma El directorio de secuencias de Proteomas
obtenidas de prodigal, el directorio de los outputs de KO y el archivo
resultado de la funcion allrast_names_table y regresa un archivo anotado con
especie y un fasta con los feature ids de rast simulados
}
\details{
This function is part of the MetaEvoMining package
}
\examples{
\dontrun{
 make_EvoFiles("inst/extdata/gtdbtk.bac120.summary.tsv",
                            "inst/extdata/Proteome_named_scaff/",
                                               "inst/KO_output/")
}
}
